Modify top20 lists to work with celery.
