+++ app/__init__.py
@@ -59,3 +59,4 @@ app.json_encoder = MyJSONEncoder
 
 from app import api, models
 from app.views import next_url, lists
+from app.tasks import *
